+++ examples/pizza_plots/plot_pizza_comparison.py
@@ -3,6 +3,8 @@
 Comparison Pizza (Percentiles)
 ==============================
 
+* Author: `slothfulwave612 <https://twitter.com/slothfulwave612>`_ 
+
 * ``mplsoccer``, ``py_pizza`` module helps one to plot pizza charts in a few lines of code.
 
 * The design idea is inspired by `Tom Worville <https://twitter.com/Worville>`_, \
@@ -110,6 +112,133 @@ fig.text(
 
 # add credits
 CREDIT_1 = "data: statsbomb viz fbref"
+CREDIT_2 = "inspired by: @Worville, @FootballSlices, @somazerofc & @Soumyaj15209314"
+
+fig.text(
+    0.99, 0.005, f"{CREDIT_1}\n{CREDIT_2}", size=9,
+    fontproperties=font_italic.prop, color="#000000",
+    ha="right"
+)
+
+plt.show()
+
+##############################################################################
+# Adjust Overlapping Values
+# -------------------------
+# To adjust overlapping values one can use ``get_compare_value_texts()`` or ``get_value_texts()``
+# to fetch a list of `axes.text` object for comparison-values-text and values-text repectively.
+# Using these method one can adjust the required text. Below is an example code.
+
+def adjust_value_text(params_offset, offset, baker):
+    """ Function to adjust the value texts. (for second player)
+    <<You can make your own function>>
+
+    Parameters
+    ----------
+    params_offset : sequence of bool, default None
+        Pass True for parameter whose value are to be adjusted
+    offset : float, default 0.0
+        Value will define how much adjustment to be made.
+    baker : PyPizza object
+    """
+    # get index value where value is True
+    idx_value = [i for i, x in enumerate(params_offset) if x]
+
+    # similarly one can use baker.get_value_texts() for first player value-texts
+    for count, (temp_text, theta) in enumerate(
+        zip(baker.get_compare_value_texts(), baker.get_theta())
+    ):
+        # fetch the adjustment-value
+        adj_val = offset if count in idx_value else 0.0
+
+        # adjust the position
+        # add some value to x-coordinate and keep y-coordinate same
+        temp_text.set_position((
+            theta+adj_val, temp_text.get_position()[1]
+        ))
+
+# parameter and values list
+params = [
+    "Non-Penalty Goals", "npxG", "npxG per Shot", "xA",
+    "Open Play\nShot Creating Actions", "\nPenalty Area\nEntries",
+    "Progressive Passes", "Progressive Carries", "Successful Dribbles",
+    "\nTouches\nper Turnover", "pAdj\nPress Regains", "Aerials Won"
+]
+
+# dummy values
+values = [15, 7, 57, 86, 63, 51, 11, 32, 85, 69, 90, 54]    # for Player 1
+values_2 = [31, 41, 43, 42, 47, 24, 60, 60, 28, 70, 92, 64]  # for Player 2
+
+# instantiate PyPizza class
+baker = PyPizza(
+    params=params,                  # list of parameters
+    background_color="#EBEBE9",     # background color
+    straight_line_color="#222222",  # color for straight lines
+    straight_line_lw=1,             # linewidth for straight lines
+    last_circle_lw=1,               # linewidth of last circle
+    last_circle_color="#222222",    # color of last circle
+    other_circle_ls="-.",           # linestyle for other circles
+    other_circle_lw=1               # linewidth for other circles
+)
+
+# plot pizza
+fig, ax = baker.make_pizza(
+    values,                     # list of values
+    compare_values=values_2,    # comparison values
+    figsize=(8, 8),             # adjust figsize according to your need
+    kwargs_slices=dict(
+        facecolor="#1A78CF", edgecolor="#222222",
+        zorder=2, linewidth=1
+    ),                          # values to be used when plotting slices
+    kwargs_compare=dict(
+        facecolor="#FF9300", edgecolor="#222222",
+        zorder=2, linewidth=1,
+    ),
+    kwargs_params=dict(
+        color="#000000", fontsize=12,
+        fontproperties=font_normal.prop, va="center"
+    ),                          # values to be used when adding parameter
+    kwargs_values=dict(
+        color="#000000", fontsize=12,
+        fontproperties=font_normal.prop, zorder=3,
+        bbox=dict(
+            edgecolor="#000000", facecolor="cornflowerblue",
+            boxstyle="round,pad=0.2", lw=1
+        )
+    ),                          # values to be used when adding parameter-values labels
+    kwargs_compare_values=dict(
+        color="#000000", fontsize=12, fontproperties=font_normal.prop, zorder=3,
+        bbox=dict(edgecolor="#000000", facecolor="#FF9300", boxstyle="round,pad=0.2", lw=1)
+    ),                          # values to be used when adding parameter-values labels
+)
+
+# pass True in that parameter-index whose values are to be adjusted
+# here True values are passed for "\nTouches\nper Turnover" and "pAdj\nPress Regains" params
+params_offset = [
+    False, False, False, False, False, False, 
+    False, False, False, True, True, False
+]
+
+# adjust the texts --> calling the function here <--
+adjust_value_text(params_offset, offset=0.13, baker=baker)
+
+# add title
+fig_text(
+    0.515, 0.99, "<Player 1> vs <Player 2>", size=17, fig=fig,
+    highlight_textprops=[{"color": '#1A78CF'}, {"color": '#EE8900'}],
+    ha="center", fontproperties=font_bold.prop, color="#000000"
+)
+
+# add subtitle
+fig.text(
+    0.515, 0.942,
+    "Percentile Rank Chart",
+    size=15,
+    ha="center", fontproperties=font_bold.prop, color="#000000"
+)
+
+# add credits 
+CREDIT_1 = "dummy-data"
 CREDIT_2 = "inspired by: @Worville, @FootballSlices, @somazerofc & @Soumyaj15209314"
 
 fig.text(